hotcook-agent/pyproject.toml
[project]
name = "hotcook-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.6.0",
    "rapidfuzz>=3.0.0",
    "claude-agent-sdk",
    "analytics-platform",
]

[tool.uv.sources]
analytics-platform = { path = "../analytics-platform" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "anyio[trio]>=4.4.0",
    "httpx>=0.27.0",
]
